[move/#5] auth 디렉토리로 이동
