+++ app/src/main/java/com/haphap/app/data/mapper/auth/AuthMapper.kt
@@ -1,7 +1,7 @@
-package com.haphap.app.data.mapper
+package com.haphap.app.data.mapper.auth
 
-import com.haphap.app.data.model.KakaoLoginModel
-import com.haphap.app.data.remote.dto.KakaoLoginResponseDto
+import com.haphap.app.data.model.auth.KakaoLoginModel
+import com.haphap.app.data.remote.dto.auth.KakaoLoginResponseDto
 
 fun KakaoLoginResponseDto.toModel(): KakaoLoginModel = KakaoLoginModel(
     accessToken = accessToken,
