+++ app/src/main/java/com/haphap/app/data/model/auth/KakaoLoginModel.kt
@@ -1,4 +1,4 @@
-package com.haphap.app.data.model
+package com.haphap.app.data.model.auth
 
 data class KakaoLoginModel(
     val accessToken: String,
