+++ app/src/main/java/com/haphap/app/data/remote/dto/auth/KakaoLoginRequestDto.kt
@@ -1,4 +1,4 @@
-package com.haphap.app.data.remote.dto
+package com.haphap.app.data.remote.dto.auth
 
 import kotlinx.serialization.Serializable
